#54 remove and/or conjunction of filters.
Instead of allowing multiple filters being connected via and/or now live filtering of the elements is possible.

Previously the user had to fill a filter, and click and/or for seeing the result of the filtering. This is now improved by typing a single character in the filter field the result is updated.

The outcome is that live filtering is possible, multiple connected filters are impossible.

Why? To simplify the search view and allow a better user experience. Less functionality but a bit more ergonomics.